stockswipe/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.4
aiokafka==0.10.0
kafka-python==2.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Data / ML
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
scikit-learn==1.4.2
torch==2.3.0
transformers==4.41.0
httpx==0.27.0
aiohttp==3.9.5

# Scheduling / jobs
apscheduler==3.10.4

# Dev / test
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0
